Move mermaid export into node test folder
