+++ tests/nodes/gr.cpp
@@ -4,7 +4,7 @@
 #include "nodes/gr/gr.h"
 #include "math/windowFunction.h"
 #include "tests/graphData.h"
-#include "tests/mermaid.h"
+#include "tests/nodes/mermaid.h"
 #include "tests/testData.h"
 #include <gtest/gtest.h>
